refactor(pneu): adiciona tipo dos campos de vida
PL-3459

+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/v3/_model/dto/PneuCadastro.java
@@ -32,10 +32,10 @@ public class PneuCadastro {
     Long codDimensao;
 
     @Nullable
-    Long vidaAtual;
+    Integer vidaAtual;
 
     @Nullable
-    Long vidaTotal;
+    Integer vidaTotal;
 
     @NotNull(message = "A pressão recomendada não pode ser nula.")
     Long pressaoRecomendada;
